+++ src/frontend-scripts/components/section-main/GamesList.jsx
@@ -4,10 +4,13 @@ import PropTypes from 'prop-types';
 import { Checkbox } from 'semantic-ui-react';
 import moment from 'moment';
 import { CURRENTSEASONNUMBER } from '../../constants';
+import { Message } from 'semantic-ui-react';
+import { processEmotes } from '../../emotes';
 
 export class GamesList extends React.Component {
 	state = {
-		filtersVisible: false
+		filtersVisible: false,
+		stickyEnabled: true
 	};
 
 	toggleFilter = value => {
@@ -15,6 +18,33 @@ export class GamesList extends React.Component {
 
 		gameFilter[value] = !gameFilter[value];
 		changeGameFilter(gameFilter);
+	};
+
+	componentWillReceiveProps(nextProps) {
+		const { generalChats } = this.props;
+		const nextGeneralChats = nextProps.generalChats;
+
+		if (!this.state.stickyEnabled && generalChats.sticky !== nextGeneralChats.sticky) {
+			this.setState({
+				stickyEnabled: true
+			});
+		}
+	}
+
+	renderSticky = () => {
+		if (this.state.stickyEnabled && this.props.generalChats.sticky) {
+			console.log(this.props.generalChats.sticky);
+			const dismissSticky = () => {
+				this.setState({ stickyEnabled: false });
+			};
+
+			return (
+				<Message onDismiss={dismissSticky} color="black">
+					Sticky:
+					{processEmotes(this.props.generalChats.sticky, true, this.props.allEmotes)}
+				</Message>
+			);
+		}
 	};
 
 	renderFilters() {
@@ -243,7 +273,10 @@ export class GamesList extends React.Component {
 				<a href="#/leaderboards" className="leaderboard">
 					Leaderboards
 				</a>
-				<div className="browser-body">{this.renderGameList()}</div>
+				<div className="browser-body">
+					{this.renderSticky()}
+					{this.renderGameList()}
+				</div>
 			</section>
 		);
 	}
@@ -262,7 +295,9 @@ GamesList.propTypes = {
 	socket: PropTypes.object,
 	userList: PropTypes.object,
 	gameFilter: PropTypes.object,
-	changeGameFilter: PropTypes.func
+	changeGameFilter: PropTypes.func,
+	generalChats: PropTypes.object,
+	allEmotes: PropTypes.object
 };
 
 export default GamesList;
