+++ src/frontend-scripts/components/section-main/gameslist.test.js
@@ -5,7 +5,8 @@ import GamesList from './GamesList';
 describe('GamesList', () => {
 	it('should initialize correctly', () => {
 		const initialState = {
-			filtersVisible: false
+			filtersVisible: false,
+			stickyEnabled: true
 		};
 
 		const component = shallow(<GamesList />);
