+++ src/scss/rightsidebar.scss
@@ -280,38 +280,6 @@
 					word-wrap: break-word;
 				}
 			}
-			.sticky {
-				position: absolute;
-				background: var(--theme-secondary);
-				top: -2px;
-				left: 0;
-				color: white;
-				padding: 3px;
-				width: 100%;
-				z-index: 1;
-				display: flex;
-				> span {
-					flex-grow: 1;
-					width: 100%;
-					word-break: break-word;
-					> span {
-						font-style: italic;
-					}
-				}
-
-				i {
-					color: white !important;
-					cursor: pointer;
-					margin: 0 3px;
-				}
-				a.shio-link {
-					color: #f4f439 !important;
-					text-decoration: underline !important;
-					&:hover {
-						color: #608cb3 !important;
-					}
-				}
-			}
 		}
 		.chats.discord {
 			height: 276px;
